Remove AgentInstrumentationConfig usages: Scheduling
