+++ instrumentation/powerjob-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/powerjob/v4_0/PowerJobSingletons.java
@@ -8,6 +8,7 @@ package io.opentelemetry.javaagent.instrumentation.powerjob.v4_0;
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.trace.StatusCode;
+import io.opentelemetry.instrumentation.api.incubator.config.internal.DeclarativeConfigUtil;
 import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesExtractor;
 import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeSpanNameExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor;
@@ -15,15 +16,15 @@ import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.instrumentation.api.instrumenter.InstrumenterBuilder;
 import io.opentelemetry.instrumentation.api.instrumenter.SpanNameExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.SpanStatusExtractor;
-import io.opentelemetry.javaagent.bootstrap.internal.AgentInstrumentationConfig;
 import tech.powerjob.worker.core.processor.ProcessResult;
 
 public final class PowerJobSingletons {
   private static final String INSTRUMENTATION_NAME = "io.opentelemetry.powerjob-4.0";
 
   private static final boolean CAPTURE_EXPERIMENTAL_SPAN_ATTRIBUTES =
-      AgentInstrumentationConfig.get()
-          .getBoolean("otel.instrumentation.powerjob.experimental-span-attributes", false);
+      DeclarativeConfigUtil.get(GlobalOpenTelemetry.get())
+          .get("powerjob")
+          .getBoolean("experimental_span_attributes", false);
   private static final Instrumenter<PowerJobProcessRequest, ProcessResult> INSTRUMENTER = create();
 
   public static Instrumenter<PowerJobProcessRequest, ProcessResult> instrumenter() {
