+++ instrumentation/quartz-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/quartz/v2_0/QuartzSingletons.java
@@ -6,16 +6,17 @@
 package io.opentelemetry.javaagent.instrumentation.quartz.v2_0;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
+import io.opentelemetry.instrumentation.api.incubator.config.internal.DeclarativeConfigUtil;
 import io.opentelemetry.instrumentation.quartz.v2_0.QuartzTelemetry;
-import io.opentelemetry.javaagent.bootstrap.internal.AgentInstrumentationConfig;
 
 public final class QuartzSingletons {
 
   public static final QuartzTelemetry TELEMETRY =
       QuartzTelemetry.builder(GlobalOpenTelemetry.get())
           .setCaptureExperimentalSpanAttributes(
-              AgentInstrumentationConfig.get()
-                  .getBoolean("otel.instrumentation.quartz.experimental-span-attributes", false))
+              DeclarativeConfigUtil.get(GlobalOpenTelemetry.get())
+                  .get("quartz")
+                  .getBoolean("experimental_span_attributes", false))
           .build();
 
   private QuartzSingletons() {}
