+++ instrumentation/xxl-job/xxl-job-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/xxljob/common/XxlJobInstrumenterFactory.java
@@ -8,17 +8,18 @@ package io.opentelemetry.javaagent.instrumentation.xxljob.common;
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.trace.StatusCode;
+import io.opentelemetry.instrumentation.api.incubator.config.internal.DeclarativeConfigUtil;
 import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.instrumentation.api.instrumenter.InstrumenterBuilder;
-import io.opentelemetry.javaagent.bootstrap.internal.AgentInstrumentationConfig;
 
 public final class XxlJobInstrumenterFactory {
 
   private static final boolean CAPTURE_EXPERIMENTAL_SPAN_ATTRIBUTES =
-      AgentInstrumentationConfig.get()
-          .getBoolean("otel.instrumentation.xxl-job.experimental-span-attributes", false);
+      DeclarativeConfigUtil.get(GlobalOpenTelemetry.get())
+          .get("xxl_job")
+          .getBoolean("experimental_span_attributes", false);
 
   public static Instrumenter<XxlJobProcessRequest, Void> create(String instrumentationName) {
     XxlJobCodeAttributesGetter codeAttributesGetter = new XxlJobCodeAttributesGetter();
